refactor: add cart class
